Add a branching parser
This adds a set of branching parsers that can consume scalars, arrays,
objects. It adds a set of API methods to the value parser provider
interface to make the parsers themselves more convenient to construct.

It also adds more unit tests, and introduces a basic set of fuzz tests
that uncovered a couple of internal exceptions and one non-termination
issue. In response to the tests, the implementations of most parsers
have been hardened.

Fix: https://github.com/irradia/one.irradia.fieldrush/issues/3